moved all redux middlewares to separate files
