+++ vis/js/middleware/queryParameter.js
@@ -1,0 +1,13 @@
+import { handleUrlAction } from "../utils/url";
+
+const queryParameterMiddleware = () => {
+  return (next) => (action) => {
+    if (!action.canceled && !action.isFromBackButton) {
+      handleUrlAction(action);
+    }
+
+    return next(action);
+  };
+};
+
+export default queryParameterMiddleware;
