+++ vis/js/middleware/repeatedInitialize.js
@@ -1,0 +1,23 @@
+/**
+ * Creates a middleware that reapplies the force layout after
+ * each additional initialization.
+ *
+ * Used in Viper rescaling
+ *
+ * @param {Object} intermediate the intermediate instance (this)
+ */
+const createRepeatedInitializeMiddleware = (forceLayoutCallback) => {
+  return ({ getState }) => {
+    return (next) => (action) => {
+      const data = getState().data.list;
+      const returnValue = next(action);
+      if (action.type === "INITIALIZE" && data.length > 0) {
+        forceLayoutCallback();
+      }
+
+      return returnValue;
+    };
+  };
+};
+
+export default createRepeatedInitializeMiddleware;
